+++ packages/devtools_app/lib/src/extension_points/extensions_external.dart
@@ -5,7 +5,7 @@
 import '../screens/debugger/codeview.dart';
 import '../shared/analytics/constants.dart' as gac;
 import '../shared/common_widgets.dart';
-import '../shared/console/eval/inspector_service.dart';
+import '../shared/inspector_service.dart';
 import '../shared/globals.dart';
 import '../shared/utils.dart';
 import 'extensions_base.dart';
