+++ packages/devtools_app/lib/src/screens/inspector/inspector_controller.dart
@@ -25,7 +25,7 @@ import '../../service/service_extensions.dart' as extensions;
 import '../../shared/config_specific/logger/logger.dart';
 import '../../shared/config_specific/url/url.dart';
 import '../../shared/console/eval/diagnostics_node.dart';
-import '../../shared/console/eval/inspector_service.dart';
+import '../../shared/inspector_service.dart';
 import '../../shared/console/eval/inspector_tree.dart';
 import '../../shared/console/primitives/instance_ref.dart';
 import '../../shared/console/primitives/simple_items.dart';
