+++ packages/devtools_app/lib/src/screens/inspector/layout_explorer/ui/layout_explorer_widget.dart
@@ -8,7 +8,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 
 import '../../../../shared/console/eval/diagnostics_node.dart';
-import '../../../../shared/console/eval/inspector_service.dart';
+import '../../../../shared/inspector_service.dart';
 import '../../../../shared/globals.dart';
 import '../../../../shared/primitives/utils.dart';
 import '../../../../shared/theme.dart';
