+++ packages/devtools_app/lib/src/screens/logging/logging_controller.dart
@@ -15,7 +15,7 @@ import 'package:vm_service/vm_service.dart';
 import '../../service/vm_service_wrapper.dart';
 import '../../shared/config_specific/logger/logger.dart' as logger;
 import '../../shared/console/eval/diagnostics_node.dart';
-import '../../shared/console/eval/inspector_service.dart';
+import '../../shared/inspector_service.dart';
 import '../../shared/console/eval/inspector_tree.dart';
 import '../../shared/globals.dart';
 import '../../shared/primitives/auto_dispose.dart';
