+++ packages/devtools_app/lib/src/screens/performance/performance_controller.dart
@@ -7,7 +7,7 @@ import 'dart:async';
 import 'package:vm_service/vm_service.dart';
 
 import '../../shared/config_specific/import_export/import_export.dart';
-import '../../shared/console/eval/inspector_service.dart';
+import '../../shared/inspector_service.dart';
 import '../../shared/feature_flags.dart';
 import '../../shared/globals.dart';
 import '../../shared/primitives/auto_dispose.dart';
