+++ packages/devtools_app/lib/src/shared/eval_on_dart_library.dart
@@ -15,7 +15,7 @@ import 'package:vm_service/vm_service.dart' as vm_service;
 
 import '../service/vm_service_wrapper.dart';
 import 'config_specific/logger/logger.dart';
-import 'console/eval/inspector_service.dart';
+import 'inspector_service.dart';
 import 'globals.dart';
 import 'primitives/auto_dispose.dart';
