+++ packages/devtools_app/lib/src/shared/inspector_service.dart
@@ -15,14 +15,14 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/foundation.dart';
 import 'package:vm_service/vm_service.dart';
 
-import '../../eval_on_dart_library.dart';
-import '../../globals.dart';
-import '../../primitives/auto_dispose.dart';
-import '../../primitives/utils.dart';
-import '../primitives/instance_ref.dart';
-import '../primitives/simple_items.dart';
-import '../primitives/source_location.dart';
-import 'diagnostics_node.dart';
+import 'eval_on_dart_library.dart';
+import 'globals.dart';
+import 'primitives/auto_dispose.dart';
+import 'primitives/utils.dart';
+import 'console/primitives/instance_ref.dart';
+import 'console/primitives/simple_items.dart';
+import 'console/primitives/source_location.dart';
+import 'console/eval/diagnostics_node.dart';
 
 const inspectorLibraryUri = 'package:flutter/src/widgets/widget_inspector.dart';
