+++ packages/devtools_app/lib/src/shared/object_tree.dart
@@ -14,7 +14,7 @@ import 'package:vm_service/vm_service.dart';
 import '../screens/debugger/debugger_model.dart';
 import 'config_specific/logger/logger.dart';
 import 'console/eval/diagnostics_node.dart';
-import 'console/eval/inspector_service.dart';
+import 'inspector_service.dart';
 import 'globals.dart';
 import 'primitives/trees.dart';
 import 'primitives/utils.dart';
